docker_MongoDB/project/tweet_collector/requirements.txt
pymongo
tweepy>=3.8.0
dnspython
python-decouple